+++ Android/NatShare/src/main/java/com/yusufolokoba/natshare/NatShare.java
@@ -11,11 +11,14 @@ import android.os.Environment;
 import android.os.StrictMode;
 import android.provider.MediaStore;
 import android.util.Log;
+import android.webkit.MimeTypeMap;
 
 import com.unity3d.player.UnityPlayer;
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.InputStream;
 import java.io.OutputStream;
 import java.nio.ByteBuffer;
 
@@ -29,6 +32,15 @@ public class NatShare {
         // Disable the FileUriExposedException from being thrown on Android 24+
         StrictMode.VmPolicy.Builder builder = new StrictMode.VmPolicy.Builder();
         StrictMode.setVmPolicy(builder.build());
+    }
+
+    public static boolean shareText (String text) {
+        final Intent intent = new Intent()
+                .setAction(Intent.ACTION_SEND)
+                .setType("text/plain")
+                .putExtra(Intent.EXTRA_TEXT, text);
+        UnityPlayer.currentActivity.startActivity(Intent.createChooser(intent, "Share"));
+        return true;
     }
 
     public static boolean shareImage (byte[] pngData, String message) {
@@ -53,12 +65,12 @@ public class NatShare {
         return true;
     }
 
-    public static boolean shareVideo (String path, String message) {
+    public static boolean shareMedia (String path, String message) {
         File file = new File(path);
         if (!file.exists()) return false;
         final Intent intent = new Intent()
                 .setAction(Intent.ACTION_SEND)
-                .setType("video/mp4")
+                .setType(getMimeType(path))
                 .addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
                 .putExtra(Intent.EXTRA_STREAM, Uri.fromFile(file))
                 .putExtra(Intent.EXTRA_TEXT, message);
@@ -66,14 +78,13 @@ public class NatShare {
         return true;
     }
 
-    public static boolean saveImageToCameraRoll (byte[] pngData) { // DEPLOY
-        ContentValues values = new ContentValues(3);
-        values.put(MediaStore.Images.Media.TITLE, "Image");
+    public static boolean saveImageToCameraRoll (byte[] pngData) {
+        final ContentValues values = new ContentValues();
         values.put(MediaStore.Images.Media.MIME_TYPE, "image/png");
-        values.put(MediaStore.Images.Media.DATE_ADDED, System.currentTimeMillis() / 1e+3);
+        values.put(MediaStore.Images.Media.DATE_ADDED, System.currentTimeMillis());
         values.put(MediaStore.Images.Media.DATE_TAKEN, System.currentTimeMillis());
-        ContentResolver resolver = UnityPlayer.currentActivity.getContentResolver();
-        Uri url = resolver.insert(MediaStore.Images.Media.EXTERNAL_CONTENT_URI, values);
+        final ContentResolver resolver = UnityPlayer.currentActivity.getContentResolver();
+        final Uri url = resolver.insert(MediaStore.Images.Media.EXTERNAL_CONTENT_URI, values);
         try {
             OutputStream stream = resolver.openOutputStream(url);
             stream.write(pngData);
@@ -87,16 +98,20 @@ public class NatShare {
         return true;
     }
 
-    public static boolean saveVideoToCameraRoll (String path) {
+    public static boolean saveMediaToCameraRoll (String path) {
         File file = new File(path);
         if (!file.exists()) return false;
-        ContentValues values = new ContentValues(3);
-        values.put(MediaStore.Video.Media.TITLE, file.getName());
-        values.put(MediaStore.Video.Media.MIME_TYPE, "video/mp4");
-        values.put(MediaStore.Images.Media.DATE_ADDED, System.currentTimeMillis() / 1e+3);
-        values.put(MediaStore.Images.Media.DATE_TAKEN, System.currentTimeMillis());
-        values.put(MediaStore.Video.Media.DATA, path);
-        UnityPlayer.currentActivity.getContentResolver().insert(MediaStore.Video.Media.EXTERNAL_CONTENT_URI, values);
+        // Copy file to gallery folder
+        File galleryDirectory = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DCIM);
+        File galleryFile = new File(galleryDirectory, file.getName());
+        try {
+            copyFile(file, galleryFile);
+        } catch (IOException ex) {
+            Log.e("Unity", "NatShare Error: Failed to copy media to camera roll");
+            ex.printStackTrace();
+            return false;
+        }
+        UnityPlayer.currentActivity.sendBroadcast(new Intent(Intent.ACTION_MEDIA_SCANNER_SCAN_FILE, Uri.fromFile(galleryFile)));
         return true;
     }
 
@@ -121,5 +136,28 @@ public class NatShare {
         frame.copyPixelsToBuffer(thumbnail.pixelBuffer);
         frame.recycle();
         return thumbnail;
+    }
+
+    private static String getMimeType (String url) {
+        String type = null;
+        String extension = MimeTypeMap.getFileExtensionFromUrl(url);
+        if (extension != null)
+            type = MimeTypeMap.getSingleton().getMimeTypeFromExtension(extension);
+        return type;
+    }
+
+    private static void copyFile (File src, File dst) throws IOException {
+        InputStream is = null;
+        OutputStream os = null;
+        try {
+            is = new FileInputStream(src);
+            os = new FileOutputStream(dst);
+            final byte[] buffer = new byte[1024];
+            int length;
+            while ((length = is.read(buffer)) > 0) os.write(buffer, 0, length);
+        } finally {
+            if (is != null) is.close();
+            if (os != null) os.close();
+        }
     }
 }